CAMEL-24570: Simplify GenAI observability no-op diagnostics
Address review feedback from Croway: log once from cached bridge and
observation backend state, use INFO consistently, remove diagnostics
helper class and doc/upgrade-guide entries, and simplify tests.
